+++ CHANGES.md
@@ -2,6 +2,15 @@ z# StreamEx changes
 
 Check also [MIGRATION.md](MIGRATION.md) for possible compatibility problems.
 
+### 2.1.2
+
+* Add `distinct(Predicate<? super Long> occurrencesFilter)`
+
+* Add `distinctBy(Function<? super T, ?> keyExtractor, Predicate<? super Long> occurrencesFilter)`.
+
+* Improvements.
+
+
 ### 2.1.1
 
 * Add `StreamEx.countBy/countByToEntry`.